feat: Home 컴포넌트에 대해 '/home'으로 리다이렉트 추가 (#5)
